Split real URL crawl tests by region
